upgrade to llama-cpp-bindings 0.11

+++ paddler_messaging/src/oversized_image_details.rs
@@ -4,6 +4,6 @@ use serde::Serialize;
 #[derive(Debug, Deserialize, Serialize)]
 #[serde(deny_unknown_fields)]
 pub struct OversizedImageDetails {
-    pub image_tokens: u32,
-    pub n_batch: u32,
+    pub image_tokens: usize,
+    pub n_batch: i32,
 }
